cpuacct: Replace fprintf() with write() to avoid fork() deadlocks.
The bionic implementation of fork() calls cpuacct_add(), which in turn
calls the stdio fopen() and fprintf().  However, these are unsafe to
call when forking a multithreaded program.  One of the discontinued
threads may have the heap locked, and calls to malloc() and free()
will cause a deadlock.  The fprintf(), fputc(), putc(), and fwrite()
are known to call malloc().

This patch replaces the stdio calls with system calls open() and
write(), which do not call malloc() or free().

This is the TI OMAP_ENHANCEMENT version of the patch.  A normal
patch has been submitted upstream to Google:

    https://partner.source.android.com/g/11853
    Change-Id: Iaf7aa19f28b262122c4f904ed596517e306cb4dc

Change-Id: Ie6b34b70945e21df2dbccc66a9d429c4c31f82d9
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>